cleanup(matrix_gen/pkg,ci): refactor matrix_gen code.
Moreover, added a build CI for it.
